+++ RELEASES.md
@@ -6,8 +6,11 @@
 
 ### Changed
  * MSR discovery now uses sysfs files instead of parsing /proc/cpuinfo
- * Message about clamping in MSR reduced from warning to informational
+ * Clamping messages reduced from warning to informational
  * Consistently reset socket count to 0 on destroy or initialization failure
+
+### Fixed
+ * Enable zones even if clamping can't be set (not available for all zones/platforms)
 
 
 ## [v0.1.2] - 2017-11-12